crawler/requirements.txt
psycopg2-binary==2.9.7
sqlalchemy==2.0.20
asyncpg==0.28.0
sqlite-utils==3.35.1
cohere>=5.0.0
pydantic==2.0.3
pydantic-settings==2.0.3
python-dotenv==1.0.0
httpx==0.24.1
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3
pgvector==0.2.3
readability-lxml==0.8.1
newspaper3k==0.2.8
schedule==1.2.0
aiohttp==3.8.5
umap-learn==0.5.3
numpy>=1.21.0
redis==5.0.1
